Rename to `isMirrored`
